fix:新增todo

+++ src/main/java/com/example/demo/study/demo2407/Demo0710.java
@@ -52,7 +52,7 @@ class calculate {
                     tempVal = tempVal * 10 + s.charAt(i) - '0';
                     i++;
                 }
-                result += flag * tempVal;
+                result += (int) (flag * tempVal);
             }
         }
